requirements.txt
django==2.2.16
djangorestframework==3.12.4
psycopg2